ビット演算にSEE2を使用。 指し手生成速度は（値の変動が激しいが）SEE4.2なしAVX2なしやねうら王ナノの3/4ぐらいが出るようになった

+++ Squirrel-2/learner.cpp
@@ -194,7 +194,7 @@ void Eval::learner()
 	timeinfo = localtime(&rawtime);
 	str = asctime(timeinfo);
 	
-	int hoge;
+	size_t hoge;
 	while ((hoge = str.find_first_of(" ")) != string::npos) {
 		str.erase(hoge, 1);
 	}
